cluster-app/scraper.py
#!/usr/bin/env python3
"""
Google Scholar Scraper and IEEE Xplore Abstract Parser.

Parses the first 15 pages of Google Scholar search results,
extracts IEEE Xplore paper links, and fetches abstracts from IEEE Xplore.
Includes robustness features like randomized delays and a mock-data fallback.
"""

import re
import time
import json
import random
import logging
import requests
from bs4 import BeautifulSoup
from urllib.parse import urljoin, urlparse, parse_qs, urlencode

logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

# Headers to mimic a modern browser request
HEADERS = {
    "User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/121.0.0.0 Safari/537.36",
    "Accept": "text/html,application/xhtml+xml,application/xml;q=0.9,image/avif,image/webp,*/*;q=0.8",
    "Accept-Language": "en-US,en;q=0.5",
    "Referer": "https://scholar.google.com/",
    "DNT": "1",
    "Connection": "keep-alive",
    "Upgrade-Insecure-Requests": "1",
}

MAX_PAGES = 15
RESULTS_PER_PAGE = 10

def parse_google_scholar(scholar_url, max_pages=MAX_PAGES):
    """
    Parse Google Scholar search results for IEEE Xplore papers.
    """
    papers = []
    session = requests.Session()
    session.headers.update(HEADERS)

    for page in range(max_pages):
        start = page * RESULTS_PER_PAGE

        parsed = urlparse(scholar_url)
        params = parse_qs(parsed.query)
        params["start"] = [str(start)]
        new_query = urlencode(params, doseq=True)
        page_url = f"{parsed.scheme}://{parsed.netloc}{parsed.path}?{new_query}"

        logger.info("Fetching Google Scholar page %d: %s", page + 1, page_url)

        try:
            response = session.get(page_url, timeout=30)
            response.raise_for_status()
        except requests.exceptions.RequestException as e:
            status_code = getattr(e.response, 'status_code', None) if e.response else None
            
            # --- MOCK MODE FALLBACK ---
            if status_code in [403, 429] or "429" in str(e) or "403" in str(e):
                logger.warning(f"Blocked by Google ({status_code})! Using sample IEEE papers to continue testing...")
                return [
                    {
                        "title": "Hadoop: A Distributed File System", 
                        "url": "https://ieeexplore.ieee.org/document/1", 
                        "ieee_id": "1", "citations": 5000, 
                        "abstract": "Hadoop is a framework that allows for the distributed processing of large data sets across clusters of computers using simple programming models. It is designed to scale up from single servers to thousands of machines, each offering local computation and storage.",
                        "authors": "Doug Cutting",
                    },
                    {
                        "title": "MapReduce: Simplified Data Processing on Large Clusters", 
                        "url": "https://ieeexplore.ieee.org/document/2", 
                        "ieee_id": "2", "citations": 12000, 
                        "abstract": "MapReduce is a programming model and an associated implementation for processing and generating large data sets. Users specify a map function that processes a key/value pair to generate a set of intermediate key/value pairs, and a reduce function that merges all intermediate values associated with the same intermediate key.",
                        "authors": "Jeffrey Dean, Sanjay Ghemawat",
                    },
                    {
                        "title": "Apache Kafka: A Distributed Streaming Platform", 
                        "url": "https://ieeexplore.ieee.org/document/3", 
                        "ieee_id": "3", "citations": 3000, 
                        "abstract": "Apache Kafka is a distributed event store and stream-processing platform. It is an open-source system developed by the Apache Software Foundation written in Java and Scala. The project aims to provide a unified, high-throughput, low-latency platform for handling real-time data feeds.",
                        "authors": "Jay Kreps",
                    }
                ]
            # --- END MOCK MODE FALLBACK ---
            
            logger.error("Failed to fetch page %d: %s", page + 1, str(e))
            continue

        soup = BeautifulSoup(response.text, "html.parser")
        results = soup.find_all("div", class_="gs_r gs_or gs_scl")
        if not results:
            results = soup.find_all("div", class_="gs_ri")

        for result in results:
            paper = extract_paper_info(result)
            if paper and paper.get("url") and "ieeexplore.ieee.org" in paper["url"]:
                papers.append(paper)
                logger.info("Found IEEE paper: %s", paper["title"][:60])

        if not soup.find("button", attrs={"aria-label": "Next"}):
            logger.info("No more pages found after page %d", page + 1)
            break

        # Randomized rate limiting to look more human-like
        sleep_time = random.uniform(5, 12)
        logger.info(f"Sleeping for {sleep_time:.2f} seconds to avoid IP block...")
        time.sleep(sleep_time)

    logger.info("Total IEEE papers found: %d", len(papers))
    return papers

# ... (rest of the file is the same) ...
