cluster-app/requirements.txt
kafka-python-ng==2.2.2
requests==2.31.0
beautifulsoup4==4.12.2
